requirements.txt
crewai
langchain
langchain-openai
python-dotenv
pandas
xlrd
openpyxl
enviPath-python
psycopg2-binary  # PostgreSQL驱动
SQLAlchemy  # 数据库工具
micom  # 用于ctFBA计算
cobra  # 用于代谢网络分析
